+++ pkgs/backend/src/entities/wms/picking_goods.rs
@@ -6,12 +6,12 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "wms_picking_goods")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub tenant_id: i32,
-    pub plan_id: i32,
-    pub sale_id: i32,
-    pub route_id: Option<i32>,
-    pub event_id: Option<i32>,
+    pub id: i64,
+    pub tenant_id: i64,
+    pub plan_id: i64,
+    pub sale_id: i64,
+    pub route_id: Option<i64>,
+    pub event_id: Option<i64>,
     pub status: i32,
     pub is_ready_to_pack: i8,
     pub created_at: DateTimeUtc,
