+++ pkgs/backend/src/entities/wms/sale_events.rs
@@ -6,10 +6,10 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "wms_sale_events")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub tenant_id: i32,
-    pub sale_id: i32,
-    pub stock_id: Option<i32>,
+    pub id: i64,
+    pub tenant_id: i64,
+    pub sale_id: i64,
+    pub stock_id: Option<i64>,
     pub status: i32,
     pub version: i32,
     pub created_at: DateTimeUtc,
